finis

+++ 0x00-python_variable_annotations/102-type_checking.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python3
 
-from typing import Tuple, List
+from typing import Tuple
 
 
-def zoom_array(lst: Tuple[int], factor: int = 2) -> Tuple[int]:
+def zoom_array(lst: Tuple[int], factor: float = 2) -> Tuple[int]:
     zoomed_in: Tuple[int] = tuple(
         item for item in lst
-        for _ in range(factor)
+        for _ in range(int(factor))
     )
     return zoomed_in
 
@@ -15,4 +15,4 @@ array = (12, 72, 91)
 
 zoom_2x = zoom_array(array)
 
-zoom_3x = zoom_array(array, 3)
+zoom_3x = zoom_array(array, 3.0)
